+++ backend/backend/views.py
@@ -9,6 +9,7 @@ from libs.user_login_throttle import UserLoginThrottle
 
 from rest_framework.views import APIView
 from substrapp.views.utils import get_channel_name
+from substrapp.orchestrator import get_orchestrator_client
 from django.conf import settings
 
 
@@ -17,12 +18,13 @@ class ExpiryObtainAuthToken(ObtainAuthToken):
     throttle_classes = [AnonRateThrottle, UserLoginThrottle]
 
     def post(self, request, *args, **kwargs):
-        serializer = self.serializer_class(data=request.data,
-                                           context={'request': request})
+        serializer = self.serializer_class(
+            data=request.data, context={"request": request}
+        )
         serializer.is_valid(raise_exception=True)
-        user = serializer.validated_data['user']
+        user = serializer.validated_data["user"]
 
-        if settings.TOKEN_STRATEGY == 'reuse':
+        if settings.TOKEN_STRATEGY == "reuse":
             token, created = Token.objects.get_or_create(user=user)
 
             # token_expire_handler will check, if the token is expired it will generate new one
@@ -33,25 +35,31 @@ class ExpiryObtainAuthToken(ObtainAuthToken):
             Token.objects.filter(user=user).delete()
             token = Token.objects.create(user=user)
 
-        return Response({
-            'token': token.key,
-            'expires_at': expires_at(token)
-        })
+        return Response({"token": token.key, "expires_at": expires_at(token)})
 
 
 class Info(APIView):
-
     def get(self, request, *args, **kwargs):
         channel_name = get_channel_name(request)
         channel = settings.LEDGER_CHANNELS[channel_name]
-        return Response({
-            'host': settings.DEFAULT_DOMAIN,
-            'version': settings.BACKEND_VERSION,
-            'channel': channel_name,
-            'config': {
-                'model_export_enabled': channel['model_export_enabled'],
-            }
-        })
+
+        with get_orchestrator_client(channel_name) as client:
+            orchestrator_versions = client.query_version()
+
+        res = {
+            "host": settings.DEFAULT_DOMAIN,
+            "version": settings.BACKEND_VERSION,
+            "orchestrator_version": orchestrator_versions.get("orchestrator"),
+            "channel": channel_name,
+            "config": {
+                "model_export_enabled": channel["model_export_enabled"],
+            },
+        }
+
+        if orchestrator_versions.get("chaincode"):
+            res["chaincode_version"] = orchestrator_versions.get("chaincode")
+
+        return Response(res)
 
 
 obtain_auth_token = ExpiryObtainAuthToken.as_view()